moved put/get/delete methods up to base class
