+++ cadc-storage-adapter-fs/src/intTest/java/org/opencadc/inventory/storage/fs/LogicalStorageAdapterTest.java
