+++ cadc-storage-adapter-fs/src/main/java/org/opencadc/inventory/storage/fs/AbstractStorageAdapter.java
@@ -67,35 +67,55 @@
 
 package org.opencadc.inventory.storage.fs;
 
+import ca.nrc.cadc.io.MultiBufferIO;
+import ca.nrc.cadc.io.ReadException;
+import ca.nrc.cadc.io.WriteException;
+import ca.nrc.cadc.net.IncorrectContentChecksumException;
+import ca.nrc.cadc.net.IncorrectContentLengthException;
+import ca.nrc.cadc.net.ResourceNotFoundException;
 import ca.nrc.cadc.net.TransientException;
 import ca.nrc.cadc.util.HexUtil;
 import ca.nrc.cadc.util.StringUtil;
 import java.io.IOException;
+import java.io.InputStream;
 import java.io.OutputStream;
 import java.io.RandomAccessFile;
 import java.net.URI;
+import java.net.URISyntaxException;
 import java.nio.ByteBuffer;
 import java.nio.charset.Charset;
 import java.nio.file.FileSystemException;
 import java.nio.file.Files;
 import java.nio.file.InvalidPathException;
 import java.nio.file.LinkOption;
+import java.nio.file.OpenOption;
 import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
 import java.nio.file.StandardOpenOption;
 import java.nio.file.attribute.UserDefinedFileAttributeView;
+import java.security.DigestOutputStream;
+import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
+import java.util.ArrayList;
 import java.util.Date;
 import java.util.Iterator;
+import java.util.List;
+import java.util.SortedSet;
+import java.util.TreeSet;
 import java.util.UUID;
 import org.apache.log4j.Logger;
 import org.opencadc.inventory.InventoryUtil;
+import org.opencadc.inventory.Namespace;
 import org.opencadc.inventory.StorageLocation;
+import org.opencadc.inventory.storage.ByteRange;
 import org.opencadc.inventory.storage.MessageDigestAPI;
+import org.opencadc.inventory.storage.NewArtifact;
 import org.opencadc.inventory.storage.PutTransaction;
 import org.opencadc.inventory.storage.StorageAdapter;
 import org.opencadc.inventory.storage.StorageEngageException;
 import org.opencadc.inventory.storage.StorageMetadata;
+import static org.opencadc.inventory.storage.fs.LogicalFileSystemStorageAdapter.CHECKSUM_ATTRIBUTE_NAME;
+import static org.opencadc.inventory.storage.fs.LogicalFileSystemStorageAdapter.MD5_CHECKSUM_SCHEME;
 import static org.opencadc.inventory.storage.fs.OpaqueFileSystemStorageAdapter.XATTR_EXEC;
 import org.opencadc.util.fs.XAttrCommandExecutor;
 
@@ -118,8 +138,15 @@ abstract class AbstractStorageAdapter implements StorageAdapter {
     static final String PREV_DIGEST_ATTR = "prevDigest";
     static final String PREV_LENGTH_ATTR = "prevLength";
     
+    private static final int CIRC_BUFFERS = 3;
+    private static final int CIRC_BUFFERSIZE = 64 * 1024;
+    static final String DELETED_PRESERVED = "deleted-preserved";
+    
     protected Path txnPath;
     protected Path contentPath;
+    
+    protected final List<Namespace> recoverableNamespaces = new ArrayList<>();
+    protected final List<Namespace> purgeNamespaces = new ArrayList<>();
     
     protected AbstractStorageAdapter() {
     }
@@ -178,14 +205,393 @@ abstract class AbstractStorageAdapter implements StorageAdapter {
     }
 
     @Override
+    public void setRecoverableNamespaces(List<Namespace> recoverable) {
+        this.recoverableNamespaces.clear();
+        this.recoverableNamespaces.addAll(recoverable);
+    }
+
+    @Override
+    public List<Namespace> getRecoverableNamespaces() {
+        return recoverableNamespaces;
+    }
+
+    @Override
+    public void setPurgeNamespaces(List<Namespace> purge) {
+        this.purgeNamespaces.clear();
+        this.purgeNamespaces.addAll(purge);
+    }
+
+    @Override
+    public List<Namespace> getPurgeNamespaces() {
+        return purgeNamespaces;
+    }
+
+    @Override
     public Iterator<PutTransaction> transactionIterator() throws StorageEngageException, TransientException {
         throw new UnsupportedOperationException();
     }
     
     @Override
+    public void get(StorageLocation storageLocation, OutputStream dest)
+        throws ResourceNotFoundException, ReadException, WriteException, StorageEngageException, TransientException {
+        InventoryUtil.assertNotNull(OpaqueFileSystemStorageAdapter.class, "storageLocation", storageLocation);
+        InventoryUtil.assertNotNull(OpaqueFileSystemStorageAdapter.class, "dest", dest);
+        log.debug("get: " + storageLocation);
+
+        Path path = storageLocationToPath(storageLocation);
+        if (!Files.exists(path)) {
+            throw new ResourceNotFoundException("not found: " + storageLocation);
+        }
+        try {
+            String delAttr = getFileAttribute(path, DELETED_PRESERVED);
+            if ("true".equals(delAttr)) {
+                log.debug("skip " + DELETED_PRESERVED + ": " + storageLocation);
+                throw new ResourceNotFoundException("not found: " + storageLocation);
+            }
+        } catch (IOException ex) {
+            throw new StorageEngageException("failed to read attributes for stored file: " + storageLocation, ex);
+        }
+        if (!Files.isRegularFile(path)) {
+            throw new IllegalArgumentException("not a file: " + storageLocation);
+        }
+        InputStream source = null;
+        try {
+            source = Files.newInputStream(path, StandardOpenOption.READ);
+        } catch (IOException ex) {
+            throw new StorageEngageException("failed to create input stream for stored file: " + storageLocation, ex);
+        }
+
+        MultiBufferIO io = new MultiBufferIO(CIRC_BUFFERS, CIRC_BUFFERSIZE);
+        try {
+            io.copy(source, dest);
+        } catch (InterruptedException ex) {
+            log.debug("get interrupted", ex);
+        }
+    }
+
+    @Override
+    public void get(StorageLocation storageLocation, OutputStream dest, ByteRange byteRange)
+        throws ResourceNotFoundException, ReadException, WriteException, StorageEngageException, TransientException {
+        InventoryUtil.assertNotNull(OpaqueFileSystemStorageAdapter.class, "storageLocation", storageLocation);
+        InventoryUtil.assertNotNull(OpaqueFileSystemStorageAdapter.class, "dest", dest);
+        InventoryUtil.assertNotNull(OpaqueFileSystemStorageAdapter.class, "byteRange", byteRange);
+        log.debug("get: " + storageLocation + " " + byteRange);
+
+        Path path = storageLocationToPath(storageLocation);
+        if (!Files.exists(path)) {
+            throw new ResourceNotFoundException("not found: " + storageLocation);
+        }
+        try {
+            String delAttr = getFileAttribute(path, DELETED_PRESERVED);
+            if ("true".equals(delAttr)) {
+                log.debug("skip " + DELETED_PRESERVED + ": " + storageLocation);
+                throw new ResourceNotFoundException("not found: " + storageLocation);
+            }
+        } catch (IOException ex) {
+            throw new StorageEngageException("failed to read attributes for stored file: " + storageLocation, ex);
+        }
+        
+        if (!Files.isRegularFile(path)) {
+            throw new IllegalArgumentException("not a file: " + storageLocation);
+        }
+        
+        InputStream source = null;
+        try {
+            if (byteRange != null) {
+                RandomAccessFile raf = new RandomAccessFile(path.toFile(), "r");
+                SortedSet<ByteRange> brs = new TreeSet<>();
+                brs.add(byteRange);
+                source = new PartialReadInputStream(raf, brs);
+            } else {
+                source = Files.newInputStream(path, StandardOpenOption.READ);
+            }
+        } catch (IOException ex) {
+            throw new StorageEngageException("failed to create input stream for stored file: " + storageLocation, ex);
+        }
+        
+        MultiBufferIO io = new MultiBufferIO(CIRC_BUFFERS, CIRC_BUFFERSIZE);
+        try {
+            io.copy(source, dest);
+        } catch (InterruptedException ex) {
+            log.debug("get interrupted", ex);
+        }
+    }
+
+    @Override
+    public StorageMetadata put(NewArtifact newArtifact, InputStream source, String transactionID)
+        throws IncorrectContentChecksumException, IncorrectContentLengthException, 
+            ReadException, WriteException,
+            StorageEngageException, TransientException {
+        InventoryUtil.assertNotNull(LogicalFileSystemStorageAdapter.class, "artifact", newArtifact);
+        InventoryUtil.assertNotNull(LogicalFileSystemStorageAdapter.class, "source", source);
+
+        Path txnTarget;
+        Long expectedLength = null;
+        MessageDigestAPI txnDigest = null;
+        String checksumAlg = DEFAULT_CHECKSUM_ALGORITHM;
+        if (newArtifact.contentChecksum != null) {
+            checksumAlg = newArtifact.contentChecksum.getScheme(); // TODO: try sha1 in here
+        }
+        try {
+            if (transactionID != null) {
+                // validate
+                UUID.fromString(transactionID);
+                txnTarget = txnPath.resolve(transactionID);
+                if (!Files.exists(txnTarget)) {
+                    throw new IllegalArgumentException("unknown transaction: " + transactionID);
+                }
+                String dstate = getFileAttribute(txnTarget, CUR_DIGEST_ATTR);
+                if (dstate != null) {
+                    txnDigest = MessageDigestAPI.getDigest(dstate);
+                }
+                if (txnDigest == null) {
+                    throw new RuntimeException("BUG: failed to restore digest in transaction " + transactionID);
+                }
+                String auri = getFileAttribute(txnTarget, ARTIFACTID_ATTR);
+                if (auri == null) {
+                    throw new RuntimeException("BUG: failed to restore uri attribute in transaction " + transactionID);
+                }
+                if (!newArtifact.getArtifactURI().toASCIIString().equals(auri)) {
+                    throw new IllegalArgumentException("incorrect Artifact.uri in transaction: " + transactionID
+                        + " expected: " + auri);
+                }
+                String alen = getFileAttribute(txnTarget, EXP_LENGTH_ATTR);
+                if (alen != null) {
+                    try {
+                        expectedLength = Long.valueOf(alen);
+                    } catch (NumberFormatException ex) {
+                        throw new RuntimeException("BUG: failed to restore expected content length attribute in transaction " + transactionID);
+                    }
+                }
+            } else {
+                String tmp = UUID.randomUUID().toString();
+                txnTarget = txnPath.resolve(tmp);
+                txnDigest = MessageDigestAPI.getInstance(checksumAlg);
+                if (Files.exists(txnTarget)) {
+                    // unlikely: duplicate UUID in the txnpath directory?
+                    throw new RuntimeException("BUG: txnTarget already exists: " + txnTarget);
+                }
+                expectedLength = newArtifact.contentLength;
+            }
+        } catch (IOException ex) {
+            throw new StorageEngageException("failed to read file attributes", ex);
+        } catch (InvalidPathException e) {
+            throw new RuntimeException("BUG: invalid path: " + txnPath, e);
+        } catch (NoSuchAlgorithmException ex) {
+            throw new RuntimeException("failed to create MessageDigestAPI: " + checksumAlg, ex);
+        }
+        log.debug("transaction: " + txnTarget + " transactionID: " + transactionID);
+        
+        
+        Throwable throwable = null;
+        URI checksum = null;
+        
+        String prevDigestState = null;
+        Long prevLength = 0L;
+        
+        try {
+            prevDigestState = MessageDigestAPI.getEncodedState(txnDigest);
+            if (transactionID != null) {
+                prevLength = Files.size(txnTarget);
+            }
+            
+            OpenOption opt = StandardOpenOption.CREATE_NEW;
+            if (transactionID != null) {
+                opt = StandardOpenOption.APPEND;
+            }
+            MessageDigestAPI md = txnDigest;
+            org.opencadc.inventory.storage.DigestOutputStream out = new org.opencadc.inventory.storage.DigestOutputStream(Files.newOutputStream(txnTarget, StandardOpenOption.WRITE, opt), txnDigest);
+            MultiBufferIO io = new MultiBufferIO();
+            if (transactionID != null) {
+                try {
+                    log.debug("append starting at offset " + prevLength);
+                    io.copy(source, out);
+                    out.flush();
+                    md = out.getMessageDigest();
+                } catch (ReadException ex) {
+                    // rollback to prevLength
+                    RandomAccessFile raf = new RandomAccessFile(txnTarget.toFile(), "rws");
+                    log.debug("rollback from " + raf.length() + " to " + prevLength + " after " + ex);
+                    raf.setLength(prevLength);
+                    md = MessageDigestAPI.getDigest(prevDigestState);
+                    long len = Files.size(txnTarget);
+                    log.debug("auto-revert transaction " + transactionID + " to length " + len + " after failed input: " + ex);
+                    throw ex;
+                } catch (WriteException ex) {
+                    log.debug("auto-abort transaction " + transactionID + " after failed write to back end: ", ex);
+                    abortTransaction(transactionID);
+                    throw ex;
+                }
+            } else {
+                io.copy(source, out);
+                out.flush();
+                md = out.getMessageDigest();
+            }
+
+            String curDigestState = MessageDigestAPI.getEncodedState(md);
+            Long curLength = Files.size(txnTarget);
+            
+            String csVal = HexUtil.toHex(md.digest());
+            checksum = URI.create(checksumAlg.toLowerCase() + ":" + csVal);
+            log.debug("current checksum: " + checksum);
+            log.debug("current file size: " + curLength);
+            
+            if (transactionID != null && (expectedLength == null || curLength < expectedLength)) {
+                // incomplete: no further content checks
+                log.debug("incomplete put in transaction: " + transactionID + " - not verifying checksum");
+            } else {
+                boolean checksumProvided = newArtifact.contentChecksum != null;
+                if (checksumProvided) {
+                    if (!newArtifact.contentChecksum.equals(checksum)) {
+                        throw new IncorrectContentChecksumException(newArtifact.contentChecksum + " != " + checksum);
+                    }
+                }
+                if (expectedLength != null && !expectedLength.equals(curLength)) {
+                    if (checksumProvided) {
+                        // likely bug in the client, throw a 400 instead
+                        throw new IncorrectContentLengthException(expectedLength + " != " + curLength
+                            + " but checksum was correct! client BUG?");
+                    }
+                    throw new IncorrectContentLengthException(expectedLength + " != " + curLength);
+                }
+            }
+
+            // Set file attributes that must be recovered in iterator
+            setFileAttribute(txnTarget, CHECKSUM_ATTR, checksum.toASCIIString());
+            setFileAttribute(txnTarget, ARTIFACTID_ATTR, newArtifact.getArtifactURI().toASCIIString());
+            
+            StorageLocation storageLocation = createStorageLocation(newArtifact.getArtifactURI(), txnTarget);
+            
+            if (transactionID != null) {
+                log.debug("transaction uncommitted: " + transactionID + " " + storageLocation);
+                // transaction will continue
+                setFileAttribute(txnTarget, CUR_DIGEST_ATTR, curDigestState);
+                setFileAttribute(txnTarget, PREV_DIGEST_ATTR, prevDigestState);
+                setFileAttribute(txnTarget, PREV_LENGTH_ATTR, prevLength.toString());
+                if (curLength == 0L) {
+                    return new StorageMetadata(storageLocation);
+                }
+                // current state
+                return new StorageMetadata(storageLocation, newArtifact.getArtifactURI(),
+                        checksum, curLength, new Date(Files.getLastModifiedTime(txnTarget).toMillis()));
+            }
+
+            // create this before committing the file so constraints applied
+            StorageMetadata metadata = new StorageMetadata(storageLocation, newArtifact.getArtifactURI(),
+                    checksum, curLength, new Date(Files.getLastModifiedTime(txnTarget).toMillis()));
+            
+            StorageMetadata ret = commit(metadata, txnTarget);
+            txnTarget = null;
+            return ret;
+            
+        } catch (ReadException | WriteException | IllegalArgumentException
+            | IncorrectContentChecksumException | IncorrectContentLengthException e) {
+            // pass through
+            throw e;
+        } catch (Throwable t) {
+            throwable = t;
+            log.error("put error", t);
+            if (throwable instanceof IOException) {
+                throw new StorageEngageException("put error", throwable);
+            }
+            // TODO: identify throwables that are transient
+            throw new RuntimeException("Unexpected error", throwable);
+        } finally {
+            // txnTarget file still exists and not in a transaction: something went wrong
+            if (txnTarget != null && transactionID == null) {
+                try {
+                    log.debug("Deleting transaction file.");
+                    Files.delete(txnTarget);
+                } catch (IOException e) {
+                    log.error("Failed to delete transaction file", e);
+                }
+            }
+        }
+    }
+ 
+@Override
+    public void delete(StorageLocation storageLocation)
+        throws ResourceNotFoundException, IOException, StorageEngageException, TransientException {
+        delete(storageLocation, false);
+    }
+    
+    @Override
+    public void delete(StorageLocation storageLocation, boolean includeRecoverable)
+        throws ResourceNotFoundException, StorageEngageException, TransientException {
+        InventoryUtil.assertNotNull(LogicalFileSystemStorageAdapter.class, "storageLocation", storageLocation);
+        Path path = storageLocationToPath(storageLocation);
+        if (!Files.exists(path)) {
+            throw new ResourceNotFoundException("not found: " + storageLocation);
+        }
+        String uriAttr = null;
+        boolean deletePreserved = false;
+        try {
+            deletePreserved = "true".equals(getFileAttribute(path, DELETED_PRESERVED));
+            uriAttr = getFileAttribute(path, ARTIFACTID_ATTR);
+        } catch (IOException ex) {
+            throw new StorageEngageException("failed to read attributes for stored file: " + storageLocation, ex);
+        }
+        
+        if (deletePreserved && !includeRecoverable) {
+            log.debug("skip " + DELETED_PRESERVED + ": " + storageLocation);
+            throw new ResourceNotFoundException("not found: " + storageLocation);
+        }
+        
+        log.debug("delete: " + storageLocation + " aka " + uriAttr);
+        if (uriAttr != null) {
+            try {
+                URI uri = new URI(uriAttr);
+                log.debug("recoverable: " + recoverableNamespaces.size() + " purge: " + purgeNamespaces.size());
+                Namespace purge = getFirstMatch(uri, purgeNamespaces);
+                Namespace recoverable = getFirstMatch(uri, recoverableNamespaces);
+               
+                if (purge != null) {
+                    log.debug("delete/purge: " + storageLocation 
+                                + " aka " + uri + " matched " + purge.getNamespace());
+                    // fall through to delete
+                } else if (recoverable != null) {
+                    log.debug("delete/recoverable: " + storageLocation 
+                                + " aka " + uri + " matched " + recoverable.getNamespace());
+                    // avoid poking fs timestamp unecessarily
+                    if (!deletePreserved) {
+                        try {
+                            setFileAttribute(path, DELETED_PRESERVED, "true");
+                        } catch (IOException ex) {
+                            throw new StorageEngageException("failed to set attribute for stored file: " + storageLocation, ex);
+                        }
+                    }
+                    return; // don't delete
+                } // else: normal fall through to delete
+            } catch (URISyntaxException ex) {
+                if (!recoverableNamespaces.isEmpty()) {
+                    log.error("found invalid " + ARTIFACTID_ATTR + "=" + uriAttr + " on " 
+                            + storageLocation + " -- cannot make recoverable");
+                }
+            }
+        } // else: no uriAttr aka incomplete put so delete
+        
+        log.debug("delete/actual: " + storageLocation + " aka " + uriAttr);
+        try {
+            Files.delete(path);
+        } catch (IOException ex) {
+            throw new StorageEngageException("failed to delete stored file: " + storageLocation, ex);
+        }
+    }
+    
+    private Namespace getFirstMatch(URI uri, List<Namespace> namespaces) {
+        for (Namespace ns : namespaces) {
+            log.debug("check: " + ns.getNamespace() + " vs " + uri);
+            if (ns.matches(uri)) {
+                return ns;
+            }
+        }
+        return null;
+    }
+
+    @Override
     public PutTransaction startTransaction(URI artifactURI, Long contentLength) 
             throws StorageEngageException, TransientException {
-        InventoryUtil.assertNotNull(FileSystemStorageAdapter.class, "artifactURI", artifactURI);
+        InventoryUtil.assertNotNull(LogicalFileSystemStorageAdapter.class, "artifactURI", artifactURI);
         try {
             String transactionID = UUID.randomUUID().toString();
             Path txnFile = txnPath.resolve(transactionID);
@@ -215,7 +621,7 @@ abstract class AbstractStorageAdapter implements StorageAdapter {
     @Override
     public PutTransaction revertTransaction(String transactionID) 
             throws IllegalArgumentException, StorageEngageException, TransientException, UnsupportedOperationException {
-        InventoryUtil.assertNotNull(FileSystemStorageAdapter.class, "transactionID", transactionID);
+        InventoryUtil.assertNotNull(LogicalFileSystemStorageAdapter.class, "transactionID", transactionID);
         try {
             // validate
             UUID.fromString(transactionID);
@@ -263,7 +669,7 @@ abstract class AbstractStorageAdapter implements StorageAdapter {
     @Override
     public void abortTransaction(String transactionID) 
             throws IllegalArgumentException, StorageEngageException, TransientException {
-        InventoryUtil.assertNotNull(FileSystemStorageAdapter.class, "transactionID", transactionID);
+        InventoryUtil.assertNotNull(LogicalFileSystemStorageAdapter.class, "transactionID", transactionID);
         try {
             Path txnTarget = txnPath.resolve(transactionID);
             if (Files.exists(txnPath)) {
@@ -279,7 +685,7 @@ abstract class AbstractStorageAdapter implements StorageAdapter {
     @Override
     public PutTransaction getTransactionStatus(String transactionID)
         throws IllegalArgumentException, StorageEngageException, TransientException {
-        InventoryUtil.assertNotNull(FileSystemStorageAdapter.class, "transactionID", transactionID);
+        InventoryUtil.assertNotNull(LogicalFileSystemStorageAdapter.class, "transactionID", transactionID);
         try {
             // validate
             UUID.fromString(transactionID);
@@ -347,6 +753,8 @@ abstract class AbstractStorageAdapter implements StorageAdapter {
         }
     }
     
+    protected abstract StorageLocation createStorageLocation(URI artifactURI, Path txnTarget);
+
     protected abstract Path storageLocationToPath(StorageLocation loc);
     
     protected abstract StorageMetadata createStorageMetadata(Path base, Path p, boolean includeRecoverable);
