+++ cadc-storage-adapter-fs/src/main/java/org/opencadc/inventory/storage/fs/LogicalIterator.java
